+++ src/libtrx/game/phase/phase_pause.c
@@ -1,5 +1,7 @@
 #include "game/phase/phase_pause.h"
 
+#include "game/console/common.h"
+#include "game/fader.h"
 #include "game/game.h"
 #include "game/game_string.h"
 #include "game/input.h"
@@ -10,12 +12,8 @@
 #include "game/shell.h"
 #include "game/sound.h"
 #include "game/text.h"
-#include "global/types.h"
-#include "global/vars.h"
-
-#include <libtrx/game/fader.h>
-#include <libtrx/game/ui/widgets/requester.h>
-#include <libtrx/memory.h>
+#include "game/ui/widgets/requester.h"
+#include "memory.h"
 
 #include <stdbool.h>
 #include <stddef.h>
@@ -33,6 +31,7 @@ typedef struct {
     bool is_ui_ready;
     UI_WIDGET *ui;
     TEXTSTRING *mode_text;
+    GAME_FLOW_ACTION action;
     FADER back_fader;
 } M_PRIV;
 
@@ -40,6 +39,8 @@ static void M_FadeIn(M_PRIV *p);
 static void M_FadeOut(M_PRIV *p);
 static void M_PauseGame(M_PRIV *p);
 static void M_ReturnToGame(M_PRIV *p);
+static void M_ExitGame(M_PRIV *p);
+static void M_ExitToTitle(M_PRIV *p);
 static void M_CreateText(M_PRIV *p);
 static void M_RemoveText(M_PRIV *p);
 static int32_t M_DisplayRequester(
@@ -52,21 +53,30 @@ static void M_Draw(PHASE *phase);
 
 static void M_FadeIn(M_PRIV *const p)
 {
+    M_CreateText(p);
     Fader_Init(&p->back_fader, FADER_TRANSPARENT, FADER_SEMI_BLACK, 0.5);
 }
 
 static void M_FadeOut(M_PRIV *const p)
 {
-    Fader_Init(&p->back_fader, FADER_ANY, FADER_TRANSPARENT, 0.3);
+    M_RemoveText(p);
+    if (p->ui != NULL) {
+        p->ui->free(p->ui);
+        p->ui = NULL;
+    }
+    if (p->action == GF_NOOP) {
+        Fader_Init(&p->back_fader, FADER_ANY, FADER_TRANSPARENT, 0.3);
+    } else {
+        Fader_Init(&p->back_fader, FADER_ANY, FADER_BLACK, 0.5);
+    }
     p->state = STATE_FADE_OUT;
 }
 
 static void M_PauseGame(M_PRIV *const p)
 {
+    p->action = GF_NOOP;
     Music_Pause();
     Sound_PauseAll();
-    Overlay_HideGameInfo();
-    M_CreateText(p);
     M_FadeIn(p);
 }
 
@@ -74,11 +84,18 @@ static void M_ReturnToGame(M_PRIV *const p)
 {
     Music_Unpause();
     Sound_UnpauseAll();
-    M_RemoveText(p);
-    if (p->ui != NULL) {
-        p->ui->free(p->ui);
-        p->ui = NULL;
-    }
+    M_FadeOut(p);
+}
+
+static void M_ExitGame(M_PRIV *const p)
+{
+    p->action = GF_EXIT_GAME;
+    M_FadeOut(p);
+}
+
+static void M_ExitToTitle(M_PRIV *const p)
+{
+    p->action = GF_EXIT_TO_TITLE;
     M_FadeOut(p);
 }
 
@@ -158,6 +175,10 @@ static PHASE_CONTROL M_Control(PHASE *const phase, int32_t const num_frames)
         p->ui->control(p->ui);
     }
 
+    if (Game_IsExiting()) {
+        M_ExitGame(p);
+    }
+
     switch (p->state) {
     case STATE_DEFAULT:
         if (g_InputDB.pause) {
@@ -184,10 +205,8 @@ static PHASE_CONTROL M_Control(PHASE *const phase, int32_t const num_frames)
         const int32_t choice = M_DisplayRequester(
             p, GS(PAUSE_ARE_YOU_SURE), GS(PAUSE_YES), GS(PAUSE_NO));
         if (choice == 0) {
-            return (PHASE_CONTROL) {
-                .action = PHASE_ACTION_END,
-                .gf_cmd = { .action = GF_EXIT_TO_TITLE },
-            };
+            M_ExitToTitle(p);
+            return (PHASE_CONTROL) { .action = PHASE_ACTION_NO_WAIT };
         } else if (choice == 1) {
             M_ReturnToGame(p);
             return (PHASE_CONTROL) { .action = PHASE_ACTION_NO_WAIT };
@@ -198,7 +217,7 @@ static PHASE_CONTROL M_Control(PHASE *const phase, int32_t const num_frames)
         if (!Fader_IsActive(&p->back_fader)) {
             return (PHASE_CONTROL) {
                 .action = PHASE_ACTION_END,
-                .gf_cmd = { .action = GF_NOOP },
+                .gf_cmd = { .action = p->action },
             };
         }
         break;
@@ -211,14 +230,18 @@ static PHASE_CONTROL M_Control(PHASE *const phase, int32_t const num_frames)
 static void M_Draw(PHASE *const phase)
 {
     M_PRIV *const p = phase->priv;
+
     Interpolation_Disable();
     Game_Draw(false);
-    Fader_Draw(&p->back_fader);
     Interpolation_Enable();
+    Fader_Draw(&p->back_fader);
+
+    Text_Draw();
     if (p->ui != NULL) {
         p->ui->draw(p->ui);
     }
-    Text_Draw();
+    Console_Draw();
+    Output_DrawPolyList();
 }
 
 PHASE *Phase_Pause_Create(void)