+++ src/libtrx/include/libtrx/game/music.h
@@ -1,0 +1,4 @@
+#pragma once
+
+extern void Music_Pause(void);
+extern void Music_Unpause(void);