+++ src/libtrx/include/libtrx/game/phase.h
@@ -2,5 +2,6 @@
 
 #include "./phase/control.h"
 #include "./phase/executor.h"
+#include "./phase/phase_pause.h"
 #include "./phase/phase_picture.h"
 #include "./phase/types.h"