+++ src/libtrx/include/libtrx/game/phase/phase_pause.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <libtrx/game/phase/types.h>
+#include "./types.h"
 
 PHASE *Phase_Pause_Create(void);
 void Phase_Pause_Destroy(PHASE *phase);