+++ src/tr1/game/phase.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include "game/phase/phase_pause.h"
 #include "game/phase/phase_photo_mode.h"
 
 #include <libtrx/game/phase.h>